Add testConcurrentModifications
